requirements.txt
# requirements.txt
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.2
ta==0.11.0
torch==2.2.2
matplotlib==3.8.3
seaborn==0.13.2
tensorboard==2.16.2
gymnasium==0.29.1
stable-baselines3==2.2.1
shimmy==0.2.1